% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{biannually_subsample}
\alias{biannually_subsample}
\title{Generate vector of biannual data}
\usage{
biannually_subsample(df)
}
\arguments{
\item{df}{A data frame with columns for Value, Date, Year, Month, Day, Quarter, Half}
}
\value{
A numeric vector
}
\description{
Generate vector of biannual data
}
\examples{
biannually_subsample(data.frame(Value = rnorm(10, 2, 0.1), Year = sample(2000:2005, 10, replace = TRUE), Month = c(1:10), Day = sample(1:30, 10)) |> dplyr::mutate(Date = paste0(Year, "-", Month, "-", Day), Half = c(rep(1, 6), rep(2, 4)), Quarter = c(rep(c(1, 2, 3), each = 3), rep(4, 1))))
}
