% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cv}
\alias{get_cv}
\title{Calculate coefficient of variation}
\usage{
get_cv(vec)
}
\arguments{
\item{vec}{a vector of numbers}
}
\value{
A number
}
\description{
Calculate coefficient of variation
}
\examples{
get_cv(c(1, 2, 1, 0, 1, 3))
}
